feat: abstract to new components
